apps/api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2
pydantic-settings>=2
backboard-sdk>=1.5
sqlalchemy>=2.0
websockets>=13
